# config/userSwagger.json
{
    "swagger" : "2.0",

    "info" : {
        "title":"User form",
        "description":"Login/Register form",
        "contact" :{
            "name":"Nitin"
        }
    },
    "produces" : [
        "application/json"
    ],
    "host":"localhost:3000",
    "basePath": "",
    "paths" : {


        "/register" : {
            "post" : {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "register",
                "parameters" :[{
                    "in" : "formData" ,
                    "name" : "name",
                    "required" : true,
                    "description" : "Your full Name ",
                    "value" : "",
                    "type" : "string"  
                },
                {
                    "in" : "formData" ,
                    "name" : "email",
                    "required" : true,
                    "description" : "Your Email id ",
                    "value" : "",
                    "type" : "string"  
                },
                {
                    "in" : "formData" ,
                    "name" : "password",
                    "required" : true,
                    "description" : "Your password",
                    "value" : "",
                    "type" : "string"  
                }
                ],
                "responses" : {}

            }
        },
        
        "/login" : {
            "post" : {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "login",
                "parameters" :[
                {
                    "in" : "formData" ,
                    "name" : "email",
                    "required" : true,
                    "description" : "Your Email id ",
                    "value" : "",
                    "type" : "string"  
                },
                {
                    "in" : "formData" ,
                    "name" : "password",
                    "required" : true,
                    "description" : "Your password",
                    "value" : "",
                    "type" : "string"  
                }
                ],
                "responses" : {}

            }
        },
        "/logout" : {
            "delete" : {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "logout",
                
                "responses" : {}

            }
        },

        "/changepassword" : {
            "post" : {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "changepassword",
                "parameters" :[{
                    "in" : "formData" ,
                    "name" : "password",
                    "required" : true,
                    "description" : "Your old password ",
                    "value" : "",
                    "type" : "string"  
                },
                {
                    "in" : "formData" ,
                    "name" : "newpassword",
                    "required" : true,
                    "description" : "Enter new password ",
                    "value" : "",
                    "type" : "string"  
                },
                {
                    "in" : "formData" ,
                    "name" : "confirmpassword",
                    "required" : true,
                    "description" : "Enter Confirm-password",
                    "value" : "",
                    "type" : "string"  
                }
                ],
                "responses" : {}

            }
        },

        "/userslist" : {
            "get" : {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "userslist",
                
                "responses" : {}

            }
        },


        "/forgotpasswordtoken" : {
            "post" : {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "changepassword",
                "parameters" :[ 
                {
                    "in" : "formData" ,
                    "name" : "email",
                    "required" : true,
                    "description" : "Enter Email",
                    "value" : "",
                    "type" : "string"  
                }
                 
                ],
                "responses" : {}

            }
        },
        "/resetpassword" : {
            "post" : {
                "x-swagger-router-controller": "middleware-name1",
                "operationId": "resetpassword",
                "parameters" :[{
                    "in" : "formData" ,
                    "name" : "token",
                    "required" : true,
                    "description" : "Your OTP ",
                    "value" : "",
                    "type" : "string"  
                },
                {
                    "in" : "formData" ,
                    "name" : "password",
                    "required" : true,
                    "description" : "Enter new password ",
                    "value" : "",
                    "type" : "string"  
                },
                {
                    "in" : "formData" ,
                    "name" : "confirmpassword",
                    "required" : true,
                    "description" : "Enter Confirm-password",
                    "value" : "",
                    "type" : "string"  
                }
                ],
                "responses" : {}

            }
        }

    }
    
}